Remove an unused cast
Seems to have been added as part of a rampant copy-paste scheme.

+++ crawl-ref/source/artefact.cc
@@ -1701,7 +1701,7 @@ static void _make_faerie_armour(item_def &item)
     item.props = doodad.props;
 
     // On body armour, an enchantment of less than 0 is never viable.
-    item.plus = max(static_cast<int>(random2(6) + random2(6) - 2), random2(2));
+    item.plus = max(random2(6) + random2(6) - 2, random2(2));
 }
 
 static jewellery_type octoring_types[8] =
